categorize components in separate dirs
