+++ root/frontend/src/components/CalendarControl/CalendarControl.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Button from './Button';
+import Button from '../Button/Button';
 
 interface PropTypes {
   incrementMonth: () => void,
