+++ root/frontend/src/routes/Home.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import CalendarPanel from '../components/CalendarPanel';
+import CalendarPanel from '../components/CalendarPanel/CalendarPanel';
 
 function Home() {
   return (
